2019/11/12作業分

+++ RocketMan/Assets/RocketMan/Components/SetButton.cs
@@ -6,7 +6,7 @@ namespace RocketMan
 {
     public struct SetButton : IComponentData
     {
-        public int3 direciton;
+        public float3 direciton;
     }
 }
